+++ src/services/index.ts
@@ -1,9 +1,5 @@
-import { FilesystemSyncher } from './filesystemSyncher';
-import { bootstrap} from './bootstrap';
 import { config } from './config';
 
 export {
-    bootstrap,
-    config,
-    FilesystemSyncher
+    config
 };
